refactor(wave4-step3): split explain behind stable facade
